pyproject.toml
[project]
name = "jobsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "certifi>=2026.2.25",
    "feedparser>=6.0.12",
    "flask>=3.1.3",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "resend>=2.26.0",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
